Fourth step of integrating DBContext and Infrastructure and Application projects to Authorization
- Isolated files in the new infrastucture projected (models and interfaces - they should be moved to the app project (which will be created)
