+++ source/marketparticipant/Energinet.DataHub.MarketParticipant.Authorization.Infrastructure/Model/ActorConsolidationActorAndDate.cs
@@ -12,16 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.MarketParticipant.Authorization.Repositories.Model;
+using System;
 
-public enum GridAreaType
+namespace Energinet.DataHub.MarketParticipant.Authorization.Infrastructure.Model;
+
+public sealed class ActorConsolidationActorAndDate
 {
-    NotSet = 0,
-    Transmission = 1,
-    Distribution = 2,
-    Other = 3,
-    Test = 4,
-    GridLossDK = 5,
-    GridLossAbroad = 6,
-    Aboard = 7
+    public Guid ActorId { get; set; }
+    public DateTimeOffset ConsolidateAt { get; set; }
 }
