+++ source/marketparticipant/Energinet.DataHub.MarketParticipant.Authorization.Infrastructure/Model/ActorName.cs
@@ -12,12 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System;
+namespace Energinet.DataHub.MarketParticipant.Authorization.Infrastructure.Model;
 
-namespace Energinet.DataHub.MarketParticipant.Authorization.Repositories.Model;
-
-public sealed class ActorConsolidationActorAndDate
-{
-    public Guid ActorId { get; set; }
-    public DateTimeOffset ConsolidateAt { get; set; }
-}
+public sealed record ActorName(string Value);
