+++ source/marketparticipant/Energinet.DataHub.MarketParticipant.Authorization.Infrastructure/Model/ActorNumber.cs
@@ -14,7 +14,7 @@
 
 using System.Text.Json.Serialization;
 
-namespace Energinet.DataHub.MarketParticipant.Authorization.Repositories.Model;
+namespace Energinet.DataHub.MarketParticipant.Authorization.Infrastructure.Model;
 
 [JsonDerivedType(typeof(EicActorNumber), typeDiscriminator: "eic")]
 [JsonDerivedType(typeof(GlnActorNumber), typeDiscriminator: "gln")]
