+++ source/marketparticipant/Energinet.DataHub.MarketParticipant.Authorization.Infrastructure/Model/ActorNumberType.cs
@@ -12,19 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System;
+namespace Energinet.DataHub.MarketParticipant.Authorization.Infrastructure.Model;
 
-namespace Energinet.DataHub.MarketParticipant.Authorization.Repositories.Model;
-
-public sealed record GridAreaOverviewItem(
-    GridAreaId Id,
-    GridAreaName Name,
-    GridAreaCode Code,
-    PriceAreaCode PriceAreaCode,
-    DateTimeOffset ValidFrom,
-    DateTimeOffset? ValidTo,
-    ActorNumber? ActorNumber,
-    ActorName? ActorName,
-    string? OrganizationName,
-    DateTimeOffset? FullFlexDate,
-    GridAreaType Type);
+public enum ActorNumberType
+{
+    Unknown = 0,
+    Gln = 1,
+    Eic = 2
+}
