+++ source/marketparticipant/Energinet.DataHub.MarketParticipant.Authorization.Infrastructure/Model/GlnActorNumber.cs
@@ -18,7 +18,7 @@ using System.Diagnostics.CodeAnalysis;
 using System.Globalization;
 using System.Linq;
 
-namespace Energinet.DataHub.MarketParticipant.Authorization.Repositories.Model;
+namespace Energinet.DataHub.MarketParticipant.Authorization.Infrastructure.Model;
 
 public sealed record GlnActorNumber : ActorNumber
 {
