+++ source/marketparticipant/Energinet.DataHub.MarketParticipant.Authorization.Infrastructure/Model/UnknownActorNumber.cs
@@ -12,8 +12,17 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System;
+using System.ComponentModel;
 
-namespace Energinet.DataHub.MarketParticipant.Authorization.Repositories.Model;
+namespace Energinet.DataHub.MarketParticipant.Authorization.Infrastructure.Model;
 
-public sealed record GridAreaId(Guid Value);
+public sealed record UnknownActorNumber : ActorNumber
+{
+    [Browsable(false)]
+    public UnknownActorNumber(string value)
+        : base(value)
+    {
+    }
+
+    public override ActorNumberType Type => ActorNumberType.Unknown;
+}
