+++ source/marketparticipant/Energinet.DataHub.MarketParticipant.Authorization.Infrastructure/Persistence/Mappers/GridAreaLinkMapper.cs
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+using Energinet.DataHub.MarketParticipant.Authorization.Infrastructure.Model;
 using Energinet.DataHub.MarketParticipant.Authorization.Infrastructure.Persistence.Model;
-using Energinet.DataHub.MarketParticipant.Authorization.Repositories.Model;
 
 namespace Energinet.DataHub.MarketParticipant.Authorization.Infrastructure.Persistence.Mappers;
