+++ source/marketparticipant/Energinet.DataHub.MarketParticipant.Authorization.Infrastructure/Persistence/Repositories/GridAreaRepository.cs
@@ -16,10 +16,10 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Energinet.DataHub.MarketParticipant.Authorization.Infrastructure.Model;
 using Energinet.DataHub.MarketParticipant.Authorization.Infrastructure.Persistence.Mappers;
 using Energinet.DataHub.MarketParticipant.Authorization.Infrastructure.Persistence.Model;
-using Energinet.DataHub.MarketParticipant.Authorization.Repositories;
-using Energinet.DataHub.MarketParticipant.Authorization.Repositories.Model;
+using Energinet.DataHub.MarketParticipant.Authorization.Infrastructure.Repositories;
 using Microsoft.EntityFrameworkCore;
 
 namespace Energinet.DataHub.MarketParticipant.Authorization.Infrastructure.Persistence.Repositories;
