+++ source/marketparticipant/Energinet.DataHub.MarketParticipant.Authorization.Infrastructure/Repositories/IGridAreaOverviewRepository.cs
@@ -12,6 +12,20 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.MarketParticipant.Authorization.Repositories.Model;
+using System.Collections.Generic;
+using System.Threading.Tasks;
+using Energinet.DataHub.MarketParticipant.Authorization.Infrastructure.Model;
 
-public sealed record ActorName(string Value);
+namespace Energinet.DataHub.MarketParticipant.Authorization.Infrastructure.Repositories;
+
+
+/// <summary>
+/// Provides access to an overview of grid areas.
+/// </summary>
+public interface IGridAreaOverviewRepository
+{
+    /// <summary>
+    /// Gets an overview of grid areas.
+    /// </summary>
+    Task<IEnumerable<GridAreaOverviewItem>> GetAsync();
+}
