+++ source/marketparticipant/Energinet.DataHub.MarketParticipant.sln
@@ -33,7 +33,7 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Energinet.DataHub.MarketPar
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Energinet.DataHub.MarketParticipant.EntryPoint.AuthApi", "Energinet.DataHub.MarketParticipant.EntryPoint.AuthApi\Energinet.DataHub.MarketParticipant.EntryPoint.AuthApi.csproj", "{757B8CD0-6BC7-43A2-CC2B-289DAF6415FC}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Energinet.DataHub.MarketParticipant.Authorization.Infrastructure", "Energinet.DataHub.MarketParticipant.Authorization.Infrastructure\Energinet.DataHub.MarketParticipant.Authorization.Infrastructure.csproj", "{8FA69340-DAA7-4415-AB57-95D3ED8F6B41}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Energinet.DataHub.MarketParticipant.Authorization.Infrastructure", "Energinet.DataHub.MarketParticipant.Authorization.Infrastructure\Energinet.DataHub.MarketParticipant.Authorization.Infrastructure.csproj", "{D50E84A1-2B91-7D26-11B5-90152C978AFA}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -93,10 +93,10 @@ Global
 		{757B8CD0-6BC7-43A2-CC2B-289DAF6415FC}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{757B8CD0-6BC7-43A2-CC2B-289DAF6415FC}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{757B8CD0-6BC7-43A2-CC2B-289DAF6415FC}.Release|Any CPU.Build.0 = Release|Any CPU
-		{8FA69340-DAA7-4415-AB57-95D3ED8F6B41}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8FA69340-DAA7-4415-AB57-95D3ED8F6B41}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8FA69340-DAA7-4415-AB57-95D3ED8F6B41}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{8FA69340-DAA7-4415-AB57-95D3ED8F6B41}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D50E84A1-2B91-7D26-11B5-90152C978AFA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D50E84A1-2B91-7D26-11B5-90152C978AFA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D50E84A1-2B91-7D26-11B5-90152C978AFA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D50E84A1-2B91-7D26-11B5-90152C978AFA}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
